Refactor delete file command and improve error handling
Updated `delete_file_command` to prevent empty list names and avoid deletion of the current list. Improved error handling for file reads in several functions to provide better diagnostic messages.

+++ src/file_management.rs
@@ -1,5 +1,6 @@
 use std::fs::{File, OpenOptions, remove_file};
 use std::{fs, io};
+use std::ffi::OsStr;
 use std::path::{ PathBuf};
 use crate::todo_struct::*;
 use std::io::{Write};
@@ -40,17 +41,21 @@ pub fn write_file(list : &Vec<Todo>, file_path : &PathBuf) -> Result<(), io::Err
 
 }
 
-pub fn delete_file(path_to_file : &PathBuf, file_name : String) {
-    let file_extension = String::from(".json");
-    let file = file_name + &file_extension;
+pub fn delete_file(path_to_file : &PathBuf, file_name : String, current_list : &PathBuf) {
+    let file_extension = ".json";
+    let file = file_name.clone() + &file_extension;
 
     let file_to_delete = path_to_file.join(file);
 
-    match remove_file(file_to_delete) {
-        Ok(()) => println!("File removed successfully"),
-        Err(e) => eprintln!("Couldn't remove the file due to {e}"),
+    if file_to_delete == OsStr::new(current_list) {
+        eprintln!("Cannot delete the current list.");
+        return;
+    } else {
+        match remove_file(file_to_delete) {
+            Ok(()) => println!("File removed successfully"),
+            Err(e) => eprintln!("Couldn't remove the file due to {e}"),
+        }
     }
-
 }
 
 pub fn auto_clean_flag(auto_clean: bool) -> bool {
